+++ src/Contracts/MessageConsumer.php
@@ -13,12 +13,15 @@ interface MessageConsumer
      */
     public function consume(): void;
 
-    /** Requests the consumer to stop after it's finished processing any messages to allow graceful exit */
-    public function stopConsume(?Closure $onStop = null): void;
+    /** Requests the consumer to stop after it's finished processing any messages to allow graceful exit. */
+    public function stopConsuming(): void;
 
     /** Will cancel the stopConsume request initiated by calling the stopConsume method */
     public function cancelStopConsume(): void;
 
     /** Count the number of messages consumed by this consumer */
     public function consumedMessagesCount(): int;
+
+    /** Defines a callable that will run when the consumer stops consuming messages. */
+    public function onStopConsuming(?Closure $onStopConsuming = null): self;
 }